--- tests/types.go
@@ -1,8 +1,0 @@
-package tests
-
-// A VMManifest contains a name and a manifest of
-// a virtual machine or virtual machine instance.
-type VMManifest struct {
-	Name     string
-	Manifest string
-}
